requirements.txt
APScheduler==3.10.4
alembic==1.16.5
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
psycopg[binary]==3.2.12
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytest==8.4.2
requests==2.32.5
SQLAlchemy==2.0.43
tzlocal==5.3.1
tzdata==2025.3
urllib3==2.6.3
